requirements.txt
flask
gunicorn
google-generativeai
dotenv
markdown2
pysqlite3